LayoutTests/fast/events/event-properties-gc-expected.txt

PASS event[prop] is "foo"
PASS event[prop] is "foo"
PASS event[prop] is "foo"
PASS successfullyParsed is true

TEST COMPLETE
